+++ attributes/forge_pack.rb
@@ -1,0 +1,4 @@
+default['forge_server']['forge_pack']['build_temp_dir'] = '/tmp/forge_server/pack_build_temp'
+
+default['forge_server']['forge_pack']['html_greeting'] = 'Modded Minecraft Resources'
+default['forge_server']['forge_pack']['location_dir'] = '/pack/'