+++ metadata.rb
@@ -7,10 +7,11 @@ long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 issues_url 'https://github.com/mrmarbury/forge_server/issues' if respond_to?(:issues_url)
 source_url 'https://github.com/mrmarbury/forge_server' if respond_to?(:source_url)
 
-version '0.0.2'
+version '1.0.4'
 
 supports 'freebsd', '>= 10.3'
 chef_version '>= 12'
 
 depends 'poise-archive', '~> 1.5.0'
 depends 'poise-monit', '~> 1.6.0'
+depends 'apache2', '~> 5.0.1'