+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG
 
-# 0.0.1
+## 0.0.1
 
 Initial Release
+
+
+## 1.0.0 
+
+Generates twitch config now
+
+## 1.0.1
+
+Fixed wrong defaultworld in 
+
+## 1.0.3
+
+providing index.html
+
+